git: commit-tree
