--- completers/git_completer/cmd/commit_tree_generated.go
@@ -1,22 +1,0 @@
-package cmd
-
-import (
-	"github.com/carapace-sh/carapace"
-	"github.com/spf13/cobra"
-)
-
-var commit_treeCmd = &cobra.Command{
-	Use:     "commit-tree",
-	Short:   "Create a new commit object",
-	Run:     func(cmd *cobra.Command, args []string) {},
-	GroupID: groups[group_low_level_manipulator].ID,
-}
-
-func init() {
-	carapace.Gen(commit_treeCmd).Standalone()
-	commit_treeCmd.Flags().StringS("F", "F", "", "read commit log message from file")
-	commit_treeCmd.Flags().StringP("gpg-sign", "S", "", "GPG sign commit")
-	commit_treeCmd.Flags().StringS("m", "m", "", "commit message")
-	commit_treeCmd.Flags().StringS("p", "p", "", "id of a parent commit object")
-	rootCmd.AddCommand(commit_treeCmd)
-}
